# Speed up addition of nullable integers with native representation