I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement

Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
- Vectorize the SUM of Nullable integer types with native representation ([David Manzanares](https://github.com/davidmanzanares), [Raúl Marín](https://github.com/Algunenano)).

Detailed description / Documentation draft:

When adding integer types, replace the null check with a multiplication (by 0 if null, by 1 if not null) to allow vectorization of the loop. [Godbolt](https://godbolt.org/z/q613adqMT). This is ~2x as fast as the original code ([Quickbench](https://quick-bench.com/q/MLTnfTvwC2qZFVeWHfOBR3U7a8I)).


## Benchmarks

Notes:
* The benchmarks are done with clickhouse-benchmark (--query) with default values.
* The tests have been done in release mode (which includes only sse4.1 and sse4.2) with and without the change.
* The PC running the test has dynamic frequency, multithreading and so on, so your mileage might vary.
* For native integers (64 bit or less) the improvement is greater the smaller the size is (I understand this is because less time is spent decompressing data, so the % adding is more) and higher at ~50% nulls (with random data this is the point were the job of the branch predictor is harder).
* Tested with Decimal32 and Decimal64: It seems to make a smaller improvement, so I've kept it here too.
* The trick wouldn't work for floating point math since the ptr could contain NaN and (NaN * 0 == NaN).
* Tested with big integers but the results were inconclusive, so it's not applied for those types. I didn't try with Decimal<BigInteger> as I expected to have similar results.
* Best case scenario: **2x** when adding uint8 with ~50% nulls.

### 1 billion u64. No nulls

```sql
CREATE TABLE nullu64_0
(
    `value` Nullable(UInt64)
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT rand32()
FROM system.numbers
LIMIT 1000000000
```

```sql
Select sum(value) from nullu64_0;
```

* Before:

`localhost:9000, queries 80, QPS: 2.862, RPS: 2862459801.186, MiB/s: 24568.690, result RPS: 2.862, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 83, QPS: 3.061, RPS: 3061115239.406, MiB/s: 26273.763, result RPS: 3.061, result MiB/s: 0.000.`

* **1.07x as fast**

### 1 billion u64. 30% nulls

```sql
CREATE TABLE nullu64_30
(
    `value` Nullable(UInt64)
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 10 < 3, NULL, v)
FROM
(
    SELECT rand32() v
    FROM system.numbers
    LIMIT 1000000000
)
```

```sql
Select sum(value) from nullu64_30;
```

* Before:

`localhost:9000, queries 99, QPS: 2.016, RPS: 2016066127.333, MiB/s: 17304.034, result RPS: 2.016, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 106, QPS: 2.582, RPS: 2582241402.968, MiB/s: 22163.556, result RPS: 2.582, result MiB/s: 0.000.`

* **1.28x as fast**


### 1 billion u64. 60% nulls


```sql
CREATE TABLE nullu64_60
(
    `value` Nullable(UInt64)
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 10 < 6, NULL, v)
FROM
(
    SELECT rand32() v
    FROM system.numbers
    LIMIT 1000000000
)
```

```sql
Select sum(value) from nullu64_60;
```

* Before:

`localhost:9000, queries 126, QPS: 1.900, RPS: 1900112804.060, MiB/s: 16308.799, result RPS: 1.900, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 103, QPS: 2.573, RPS: 2572793393.127, MiB/s: 22082.463, result RPS: 2.573, result MiB/s: 0.000.`

* **1.35x as fast**

### 1 billion u64. 99% nulls

```sql
CREATE TABLE nullu64_99
(
    `value` Nullable(UInt64)
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 100 < 99, NULL, v)
FROM
(
    SELECT rand32() v
    FROM system.numbers
    LIMIT 1000000000
)
```

```sql
Select sum(value) from nullu64_99;
```

* Before:

`localhost:9000, queries 114, QPS: 2.809, RPS: 2809414834.123, MiB/s: 24113.401, result RPS: 2.809, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 109, QPS: 2.948, RPS: 2948173491.255, MiB/s: 25304.376, result RPS: 2.948, result MiB/s: 0.000.`

* **1.04x as fast**

### 1 billion i16. 50% nulls


```sql
CREATE TABLE null16_50
(
    `value` Nullable(Int16)
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 10 < 5, NULL, v)
FROM
(
    SELECT rand32() % (pow(2,15)) v
    FROM system.numbers
    LIMIT 1000000000
)
```


```sql
Select sum(value) from null16_50;
```

* Before:

`localhost:9000, queries 107, QPS: 3.128, RPS: 3127769193.658, MiB/s: 8948.619, result RPS: 3.128, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 111, QPS: 5.084, RPS: 5083888739.705, MiB/s: 14545.122, result RPS: 5.084, result MiB/s: 0.000.`

* **1.62x as fast**

### 1 billion i8. 50% nulls


```sql
CREATE TABLE nullu8_50
(
    `value` Nullable(UInt8)
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 10 < 5, NULL, v)
FROM
(
    SELECT rand32() % (pow(2,8)) v
    FROM system.numbers
    LIMIT 1000000000
)
```


```sql
Select sum(value) from nullu8_50;
```

* Before:

`localhost:9000, queries 104, QPS: 3.272, RPS: 3272032259.383, MiB/s: 6240.906, result RPS: 3.272, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 126, QPS: 6.526, RPS: 6525698468.953, MiB/s: 12446.782, result RPS: 6.526, result MiB/s: 0.000.`

* **1.99x as fast**


### 1 billion Decimal 32. 50% nulls


```sql
CREATE TABLE nulld32_50
(
    `value` Nullable(Decimal32(4))
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 10 < 5, NULL, v)
FROM
(
    SELECT rand32() % (pow(10,5)) v
    FROM system.numbers
    LIMIT 1000000000
)
```


```sql
Select sum(value) from nulld32_50;
```

* Before:

`localhost:9000, queries 107, QPS: 2.593, RPS: 2592640253.900, MiB/s: 12362.672, result RPS: 2.593, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 105, QPS: 2.772, RPS: 2772381444.817, MiB/s: 13219.745, result RPS: 2.772, result MiB/s: 0.000.`

* **1.07x as fast**


### 1 billion Decimal 64. 50% nulls


```sql
CREATE TABLE nulld64_50
(
    `value` Nullable(Decimal64(8))
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 10 < 5, NULL, v)
FROM
(
    SELECT rand32() % (pow(10,10)) v
    FROM system.numbers
    LIMIT 1000000000
)
```


```sql
Select sum(value) from nulld64_50;
```

* Before:

`localhost:9000, queries 206, QPS: 1.847, RPS: 1847488808.696, MiB/s: 15857.124, result RPS: 1.847, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 107, QPS: 1.922, RPS: 1922299170.575, MiB/s: 16499.226, result RPS: 1.922, result MiB/s: 0.000.`


* **1.04x as fast**


## OTHER TYPES (Discarded)


### 1 billion i128. 50% nulls (Discarded)

```sql
CREATE TABLE null128_50
(
    `value` Nullable(Int128)
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 10 < 5, NULL, v)
FROM
(
    SELECT rand64() v
    FROM system.numbers
    LIMIT 1000000000
)
```

```sql
Select sum(value) from null128_50;
```

* Before:

`localhost:9000, queries 102, QPS: 1.480, RPS: 1480322701.877, MiB/s: 23999.678, result RPS: 1.480, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 105, QPS: 1.563, RPS: 1563451585.408, MiB/s: 25347.402, result RPS: 1.563, result MiB/s: 0.000.`

* 1.05 as fast


### 1 billion u256. 50% nulls (Discarded)

```sql
CREATE TABLE nullu256_50
(
    `value` Nullable(UInt256)
)
ENGINE = MergeTree
ORDER BY tuple() AS
SELECT IF(v % 10 < 5, NULL, v)
FROM
(
    SELECT rand64() v
    FROM system.numbers
    LIMIT 1000000000
)
```


```sql
Select sum(value) from nullu256_50;
```

* Before:

`localhost:9000, queries 222, QPS: 1.195, RPS: 1195412714.886, MiB/s: 37621.135, result RPS: 1.195, result MiB/s: 0.000.`

* After:

`localhost:9000, queries 102, QPS: 1.191, RPS: 1191439794.415, MiB/s: 37496.103, result RPS: 1.191, result MiB/s: 0.000.`

* 0.99 as fast